move nlp to separate folder
