+++ app/views.py
@@ -11,54 +11,6 @@ from forms import ExampleForm, LoginForm
 from models import User
 import six
 
-# Imports the Google Cloud client library
-from google.cloud import language
-from google.cloud.language import enums
-from google.cloud.language import types
-
-def entities_text(text):
-    """Detects entities in the text."""
-    client = language.LanguageServiceClient()
-
-    if isinstance(text, six.binary_type):
-        text = text.decode('utf-8')
-
-    # Instantiates a plain text document.
-    document = types.Document(
-        content=text,
-        type=enums.Document.Type.PLAIN_TEXT)
-
-    # Detects entities in the document. You can also analyze HTML with:
-    #   document.type == enums.Document.Type.HTML
-    entities = client.analyze_entities(document).entities
-
-    # entity types from enums.Entity.Type
-    entity_type = ('UNKNOWN', 'PERSON', 'LOCATION', 'ORGANIZATION',
-                   'EVENT', 'WORK_OF_ART', 'CONSUMER_GOOD', 'OTHER')
-
-	# Detects syntax in the document. You can also analyze HTML with:
-    #   document.type == enums.Document.Type.HTML
-    tokens = client.analyze_syntax(document).tokens
-
-    # part-of-speech tags from enums.PartOfSpeech.Tag
-    pos_tag = ('UNKNOWN', 'ADJ', 'ADP', 'ADV', 'CONJ', 'DET', 'NOUN', 'NUM',
-               'PRON', 'PRT', 'PUNCT', 'VERB', 'X', 'AFFIX')
-
-    for token in tokens:
-        print(u'{}: {}'.format(pos_tag[token.part_of_speech.tag],
-                               token.text.content))
-
-    for entity in entities:
-        print('=' * 20)
-        print(u'{:<16}: {}'.format('name', entity.name))
-        print(u'{:<16}: {}'.format('type', entity_type[entity.type]))
-        print(u'{:<16}: {}'.format('metadata', entity.metadata))
-        print(u'{:<16}: {}'.format('salience', entity.salience))
-        print(u'{:<16}: {}'.format('wikipedia_url',
-              entity.metadata.get('wikipedia_url', '-')))
-
-entities_text('Would you like a can of soda')
-
 @app.route('/')
 def index():
 	return render_template('index.html')
